1. 기본DB를 H2에서 Mysql로 변경 2. 게시글목록조회 개발(작업중)

+++ src/main/java/kr/co/puerpuella/apitextssul/api/article/dto/request/Art01Request.java
@@ -14,23 +14,23 @@ import java.util.List;
 public class Art01Request extends CommonDTO {
 
     /** 검색 시작 위치 */
-    private int offset = 0;
+    private Integer offset = 0;
 
     /** 정렬방식(asc, desc) */
     private String order = "asc";
 
     /** 표시 수 */
-    private int limit = 20;
+    private Integer limit = 20;
 
     /** 검색조건(작성자) */
-    private int autorUid;
+    private Long autorUid;
 
     /** 검색조건(제목) */
     private String articleTitle;
 
     /** 검색조건(카테고리) */
-    private int categoryId;
+    private Integer categoryId;
 
     /** 검색조건(제목) */
-    private int articleTypeId;
+    private Integer articleTypeId;
 }
